Fix menu mobile layout
